HOMIFI_PROJECT/requirements.txt
Django>=5.1.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
mysqlclient>=2.2.0
Pillow>=10.1.0  # For image handling
python-dotenv>=1.0.0  # For environment variables
requests>=2.31.0  # For Google Maps API calls